vis/requirements.txt
PyOpenGL==3.1.5
freetype-py~=2.2.0
numpy~=1.22.1
PyGLM~-2.5.7